Implemented concurrent workers
