+++ tests/Network/Nakadi/Examples/Subscription/Test.hs
@@ -44,9 +44,11 @@ testConsumption config = Nakadi.runNakadiT config $ do
   bracket before after $ \ subscriptionId -> do
     _ <- flip runLoggingT (logger nakadiLogRef) $ do
       events <- genEvents
-      void . async $ delayedPublish Nothing events
-      withAsync (dumpSubscription subscriptionId) $ \ _dumpHandle ->
-        threadDelay (2 * 10^6) -- Give Nakadi some time to transmit the published events
+      publisherHandle <- async $ delayedPublish Nothing events
+      link publisherHandle
+      withAsync (dumpSubscription subscriptionId) $ \ dumpHandle -> do
+        link dumpHandle
+        threadDelay (5 * 10^6) -- Give Nakadi some time to transmit the published events
     nakadiLog <- liftIO $ readIORef nakadiLogRef
     when (length nakadiLog == 0) $
       liftIO $ assertFailure "Subscription Consumption has logged no received batches"
